Python: Move PubSubMsg into a shared file
