+++ python/python/glide/__init__.py
@@ -1,6 +1,5 @@
 # Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0
 
-from glide.commands.async_commands.core import CoreCommands
 from glide.commands.batch import (
     Batch,
     ClusterBatch,
@@ -38,6 +37,7 @@ from glide.commands.core_options import (
     InfoSection,
     InsertPosition,
     OnlyIfEqual,
+    PubSubMsg,
     UpdateOptions,
 )
 from glide.commands.server_modules import ft, glide_json, json_batch
@@ -169,8 +169,6 @@ from glide.routes import (
 )
 
 from .glide import ClusterScanCursor, Script
-
-PubSubMsg = CoreCommands.PubSubMsg
 
 __all__ = [
     # Client
