+++ python/python/glide/config.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 from enum import Enum, IntEnum
 from typing import Any, Callable, Dict, List, Optional, Set, Tuple, Union
 
-from glide.commands.async_commands.core import CoreCommands
+from glide.commands.core_options import PubSubMsg
 from glide.exceptions import ConfigurationError
 from glide.protobuf.connection_request_pb2 import ConnectionRequest
 from glide.protobuf.connection_request_pb2 import ProtocolVersion as SentProtocolVersion
@@ -308,7 +308,7 @@ class BaseClientConfiguration:
 
     def _get_pubsub_callback_and_context(
         self,
-    ) -> Tuple[Optional[Callable[[CoreCommands.PubSubMsg, Any], None]], Any]:
+    ) -> Tuple[Optional[Callable[[PubSubMsg, Any], None]], Any]:
         return None, None
 
 
@@ -387,7 +387,7 @@ class GlideClientConfiguration(BaseClientConfiguration):
         Attributes:
             channels_and_patterns (Dict[GlideClientConfiguration.PubSubChannelModes, Set[str]]):
                 Channels and patterns by modes.
-            callback (Optional[Callable[[CoreCommands.PubSubMsg, Any], None]]):
+            callback (Optional[Callable[[PubSubMsg, Any], None]]):
                 Optional callback to accept the incomming messages.
             context (Any):
                 Arbitrary context to pass to the callback.
@@ -396,7 +396,7 @@ class GlideClientConfiguration(BaseClientConfiguration):
         channels_and_patterns: Dict[
             GlideClientConfiguration.PubSubChannelModes, Set[str]
         ]
-        callback: Optional[Callable[[CoreCommands.PubSubMsg, Any], None]]
+        callback: Optional[Callable[[PubSubMsg, Any], None]]
         context: Any
 
     def __init__(
@@ -468,7 +468,7 @@ class GlideClientConfiguration(BaseClientConfiguration):
 
     def _get_pubsub_callback_and_context(
         self,
-    ) -> Tuple[Optional[Callable[[CoreCommands.PubSubMsg, Any], None]], Any]:
+    ) -> Tuple[Optional[Callable[[PubSubMsg, Any], None]], Any]:
         if self.pubsub_subscriptions:
             return self.pubsub_subscriptions.callback, self.pubsub_subscriptions.context
         return None, None
@@ -557,7 +557,7 @@ class GlideClusterClientConfiguration(BaseClientConfiguration):
         Attributes:
             channels_and_patterns (Dict[GlideClusterClientConfiguration.PubSubChannelModes, Set[str]]):
                 Channels and patterns by modes.
-            callback (Optional[Callable[[CoreCommands.PubSubMsg, Any], None]]):
+            callback (Optional[Callable[[PubSubMsg, Any], None]]):
                 Optional callback to accept the incoming messages.
             context (Any):
                 Arbitrary context to pass to the callback.
@@ -566,7 +566,7 @@ class GlideClusterClientConfiguration(BaseClientConfiguration):
         channels_and_patterns: Dict[
             GlideClusterClientConfiguration.PubSubChannelModes, Set[str]
         ]
-        callback: Optional[Callable[[CoreCommands.PubSubMsg, Any], None]]
+        callback: Optional[Callable[[PubSubMsg, Any], None]]
         context: Any
 
     def __init__(
@@ -644,7 +644,7 @@ class GlideClusterClientConfiguration(BaseClientConfiguration):
 
     def _get_pubsub_callback_and_context(
         self,
-    ) -> Tuple[Optional[Callable[[CoreCommands.PubSubMsg, Any], None]], Any]:
+    ) -> Tuple[Optional[Callable[[PubSubMsg, Any], None]], Any]:
         if self.pubsub_subscriptions:
             return self.pubsub_subscriptions.callback, self.pubsub_subscriptions.context
         return None, None
